Add focusable buttons to DOM
